Create LoginLocationAnalysis
This is an updated query that retrieves Azure sign in location data and compares the users last signin location. Providing the analyst with a data set that shows login start time, user details, location, IP address & difference in location by miles. You can use this to easily distinguish if a user is logging in from a new location/IP.